packages/toolbox-core/pyproject.toml
[project]
name = "toolbox-core"
dynamic = ["version"]
readme = "README.md"
description = "Python Base SDK for interacting with the Toolbox service"
license = {file = "LICENSE"}
requires-python = ">=3.9"
authors = [
    {name = "Google LLC", email = "googleapis-packages@google.com"}
]

dependencies = [
    "pydantic>=2.7.0,<3.0.0",
    "aiohttp>=3.8.6,<4.0.0",
]

classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

# Tells setuptools that packages are under the 'src' directory
[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.dynamic]
version = {attr = "toolbox_core.version.__version__"}

[project.urls]
Homepage = "https://github.com/googleapis/mcp-toolbox-sdk-python/blob/main/packages/toolbox-core"
Repository = "https://github.com/googleapis/mcp-toolbox-sdk-python.git"
"Bug Tracker" = "https://github.com/googleapis/mcp-toolbox-sdk-python/issues"
Changelog = "https://github.com/googleapis/mcp-toolbox-sdk-python/blob/main/packages/toolbox-core/CHANGELOG.md"

[project.optional-dependencies]
test = [
    "black[jupyter]==25.1.0",
    "isort==6.0.1",
    "mypy==1.15.0",
    "pytest==8.3.5",
    "pytest-aioresponses==0.3.0",
    "pytest-asyncio==0.26.0",
    "pytest-cov==6.1.0",
    "google-cloud-secret-manager==2.23.2",
    "google-cloud-storage==3.1.0",
]
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.black]
target-version = ['py39']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.9"
warn_unused_configs = true
disallow_incomplete_defs = true
